+++ changelog.md
@@ -1,5 +1,12 @@
 ## Changelog
 
+### 9.21.0
+
+* `[system/process]` Added new type `ProcSample` as a lightweight analog of ProcInfo for CPU usage calculation
+* `[system/process]` Code refactoring
+* `[system/process]` Increased code coverage (75.5% → 82.4%)
+* `[system]` Code refactoring
+
 ### 9.20.1
 
 * `[fmtutil]` Added method `PrettyPerc` for formatting values in percentages